+++ themes/2003_v1/default_template.php
@@ -1,10 +1,7 @@
-<?php
-$page_title = $page_title ?? 'Steam Powered';
-include $CMS_ROOT.'/header.php';
-?>
+{header}
 <div align="center">
 <table bgcolor="#4c5844" border="0" cellpadding="0" cellspacing="0" width="800">
 <tr><td>{content}</td></tr>
 </table>
 </div>
-<?php include $CMS_ROOT.'/footer.php'; ?>
+{footer}