+++ themes/2003_v1/index_template.php
@@ -1,6 +1,3 @@
-<?php
-$page_title = $page_title ?? 'Welcome to Steam';
-include $CMS_ROOT.'/header.php';
-?>
+{header}
 {content}
-<?php include $CMS_ROOT.'/footer.php'; ?>
+{footer}