+++ themes/2003_v2/default_template.php
@@ -1,7 +1,4 @@
-<?php
-$page_title = $page_title ?? 'Steam';
-include $CMS_ROOT.'/header.php';
-?>
+{header}
 <div class="content" id="container">
 <h1>TITLE</h1>
 <h2>basic saying, <em>about page</em></h2><img src="<?php echo htmlspecialchars($THEME_URL); ?>/img/Graphic_box.jpg" height="6" width="24" alt=""><br>
@@ -9,4 +6,4 @@ include $CMS_ROOT.'/header.php';
 {content}
 </div>
 </div>
-<?php include $CMS_ROOT.'/footer.php'; ?>
+{footer}