+++ themes/2004/default_template.php
@@ -1,8 +1,5 @@
-<?php
-$page_title = 'PAGE TITLE';
-include $CMS_ROOT.'/header.php';
-?>
+{header}
 <div class="content" id="container">
 {content}
 </div>
-<?php include $CMS_ROOT.'/footer.php'; ?>
+{footer}