+++ themes/2004/footer.php
@@ -1,8 +1,0 @@
-<?php
-$root = dirname(__DIR__,2);
-require_once "$root/cms/db.php";
-$base = cms_base_url();
-echo cms_get_theme_footer('2004');
-?>
-</body>
-</html>